move dag logic to utils folder
