+++ .gitignore
@@ -3,3 +3,4 @@ dist/
 sopic.egg-info/
 __pycache__/
 .pytest_cache/
+dag-*.png
