+++ sopic/utils/__init__.py
@@ -1,2 +1,1 @@
 from .logger import getLogger, getDefaultFormatter, getColorFormatter
-from .settings import overwrite_settings_values, step_settings
